+++ packages/react-auth/src/AuthProvider/index.tsx
@@ -6,6 +6,7 @@ import {
   useContext,
   useEffect,
   useMemo,
+  useRef,
   useState,
 } from "react";
 
@@ -15,16 +16,19 @@ import {
   getStorageAPI,
   getSessionData,
   getSessionToken,
-  getRefreshToken,
+  getAllTokens,
+  getTokenCookieFields,
   getUserFromResponse,
   hasAuthParams,
   saveSessionData,
   setTokenCookies,
   removeTokenCookies,
-  startTokenWatch,
+  getTokenExpire,
+  isTokenExpiring,
   getTokenFromCookie,
   SESSION_DATA_KEY,
   DEFAULT_COOKIE_OPTIONS,
+  REFRESH_CHECK_INTERVAL,
 } from "@src/shared/session";
 import {
   APIResponse,
@@ -141,7 +145,7 @@ export const AuthProvider: FC<AuthProviderProps> = ({
   const [loading, setLoading] = useState<boolean>(true);
   const [error, setError] = useState<string>("");
   const [user, setUser] = useState<AuthUser>();
-  const [timer, setTimer] = useState<number>();
+  const intervalId = useRef<number | null>(null);
 
   const client = useMemo(() => {
     return new AuthNClient(config);
@@ -162,52 +166,59 @@ export const AuthProvider: FC<AuthProviderProps> = ({
   };
 
   useEffect(() => {
-    const storageAPI = getStorageAPI(useCookie);
-    const token = getSessionToken(options);
+    const [token, expire] = getTokenCookieFields(options.cookieName as string);
 
     if (hasAuthParams()) {
       // if code and secret params are set, exchange code for a token
       exchange();
     } else if (token) {
-      // if token exists, check if it's valid
-      validate(token);
+      // if token is expiring or expired, try refreshing
+      if (expire && isTokenExpiring(expire)) {
+        refresh();
+      } else {
+        // if token has not expired, validate that it's still good
+        validate(token);
+      }
+
+      startTokenWatch();
     } else {
       // show unauthenticated state
       setLoading(false);
     }
 
-    // clear the timer on unmount, if it's set
+    // event handler to start/stop refresh checker
+    document.addEventListener("visibilitychange", checkVisibility);
+
     return () => {
-      if (timer) {
-        clearInterval(timer);
+      document.removeEventListener("visibilitychange", checkVisibility);
+      if (intervalId.current) {
+        clearInterval(intervalId.current);
+        intervalId.current = null;
       }
     };
   }, []);
 
-  const validate = async (token: string) => {
-    const refreshToken = getRefreshToken(options);
+  const checkVisibility = () => {
+    if (document.hidden) {
+      if (intervalId.current) {
+        clearInterval(intervalId.current);
+        intervalId.current = null;
+      }
+    } else {
+      setLoading(true);
+      checkTokenLife();
+      setLoading(false);
+      startTokenWatch();
+    }
+  };
 
-    const { success, response } = refreshToken
-      ? await client.refresh(token, refreshToken)
-      : await client.validate(token);
+  const validate = async (token: string) => {
+    const { success, response } = await client.validate(token);
 
     if (success) {
-      const user: AuthUser = getUserFromResponse(response);
       const sessionData = getSessionData(options);
-      sessionData.user = user;
 
-      saveSessionData(sessionData, options);
-      setUser(sessionData.user);
       setAuthenticated(true);
-
-      if (useCookie) {
-        setTokenCookies(user, options);
-      }
-
-      const timerId = startTokenWatch(refresh, options);
-      if (timerId) {
-        setTimer(timerId);
-      }
 
       if (onLogin) {
         const appState = {
@@ -314,7 +325,7 @@ export const AuthProvider: FC<AuthProviderProps> = ({
     }
     // call the logout endpoint
     else {
-      const userToken = user?.active_token.token;
+      const userToken = getSessionToken(options);
 
       if (userToken) {
         const { success, response } = await client.logout(userToken);
@@ -334,16 +345,16 @@ export const AuthProvider: FC<AuthProviderProps> = ({
     return getSessionToken(options);
   }, []);
 
-  const refresh = async (useCookie: boolean) => {
-    const sessionData = getSessionData(options);
-    const activeToken = sessionData.user?.active_token?.token || "";
-    const refreshToken = sessionData.user?.refresh_token?.token || "";
+  const refresh = async () => {
+    const { sessionToken, refreshToken } = getAllTokens(options);
+
     const { success, response } = await client.refresh(
-      activeToken,
+      sessionToken,
       refreshToken
     );
 
     if (success) {
+      const sessionData = getSessionData(options);
       const user: AuthUser = getUserFromResponse(response);
       sessionData.user = user;
       saveSessionData(sessionData, options);
@@ -352,6 +363,9 @@ export const AuthProvider: FC<AuthProviderProps> = ({
       if (useCookie) {
         setTokenCookies(user, options);
       }
+
+      setAuthenticated(true);
+      setLoading(false);
     } else {
       logout();
     }
@@ -362,8 +376,9 @@ export const AuthProvider: FC<AuthProviderProps> = ({
       removeTokenCookies(options);
     }
 
-    if (timer) {
-      clearInterval(timer);
+    if (intervalId.current) {
+      clearInterval(intervalId.current);
+      intervalId.current = null;
     }
 
     const storageAPI = getStorageAPI(useCookie);
@@ -399,14 +414,31 @@ export const AuthProvider: FC<AuthProviderProps> = ({
     setUser(user);
     setAuthenticated(true);
 
-    const timerId = startTokenWatch(refresh, options);
-    if (timerId) {
-      setTimer(timerId);
-    }
+    startTokenWatch();
 
     if (onLogin) {
       onLogin(appState);
     }
+  };
+
+  const checkTokenLife = () => {
+    const tokenExpire = getTokenExpire(options);
+    if (tokenExpire && isTokenExpiring(tokenExpire)) {
+      refresh();
+    }
+  };
+
+  const startTokenWatch = () => {
+    const intervalTime = REFRESH_CHECK_INTERVAL * 1000;
+
+    if (intervalId.current) {
+      clearInterval(intervalId.current);
+      intervalId.current = null;
+    }
+
+    intervalId.current = window.setInterval(() => {
+      checkTokenLife();
+    }, intervalTime);
   };
 
   const memoData = useMemo(