+++ packages/react-mui-authn/src/views/Otp/index.tsx
@@ -6,13 +6,18 @@ import { ThemeOptions } from "@mui/material/styles";
 import AuthNPanel from "@src/components/core/Panel";
 import CodeForm from "@src/components/forms/CodeForm";
 
-interface OtpViewProps {
+interface OtpViewComponentProps {
   config: any; // TODO: add shared interface
   themeOptions?: ThemeOptions;
   sx?: SxProps;
 }
 
-const OtpView: FC<OtpViewProps> = ({ config, themeOptions, sx, ...props }) => {
+const OtpView: FC<OtpViewComponentProps> = ({
+  config,
+  themeOptions,
+  sx,
+  ...props
+}) => {
   const formTitle = "Let's verify it's you";
   const bodyContent =
     "A six digit code was sent to *******5309. Enter the code below.";